Updated branch plus 'throw detection'
Updated the branch to latest version of master. Holding an object requires the button to be held down again (temp). Pressing R while holding a cube now drops the cube and makes it unable to be picked up for a small duration.

Need to find physics bug causing cubes to disappear/move very fast.